requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
celery==5.3.4
redis==5.0.1
mysqlclient==2.2.0
django-environ==0.10.0
Pillow==10.0.1
python-decouple==3.8